pyproject.toml
[project]
name = "lfi3751-to-influxdb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "influxdb-client>=1.50.0",
    "pyserial>=3.5",
    "pytest>=9.0.2",
]
